ci: bump actions versions
The original file got outdated: the deployment fails with the following error

"This request has been automatically failed because it uses a deprecated version of `actions/upload-artifact: v3`. Learn more: https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/ "

This might confuse new users, so I bumped the versions of upload-artifact and other actions. It all seems to work now, but I'm no expert in GitHub Actions so it might be good to re-check me